requirements.txt
torch
clip
Pillow
gradio
numpy
scikit-learn
python-dotenv
pysmb
